# internal/api/tags.go
// Governing: SPEC-0005 REQ "Tags", ADR-0008
package api

import (
	"errors"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/joestump/joe-links/internal/auth"
	"github.com/joestump/joe-links/internal/store"
)

// tagsAPIHandler provides REST handlers for tag endpoints.
// Governing: SPEC-0005 REQ "Tags"
type tagsAPIHandler struct {
	tags  *store.TagStore
	links *store.LinkStore
}

// registerTagRoutes registers tag routes on r.
// Governing: SPEC-0005 REQ "Tags"
func registerTagRoutes(r chi.Router, tags *store.TagStore, links *store.LinkStore) {
	h := &tagsAPIHandler{tags: tags, links: links}
	r.Get("/tags", h.List)
	r.Get("/tags/{slug}/links", h.ListLinks)
}

// List returns all tags with link_count >= 1.
// GET /api/v1/tags
// Governing: SPEC-0005 REQ "Tags" — tags with link_count = 0 MUST NOT appear.
//
// @Summary      List tags
// @Description  Returns all tags that have at least one associated link.
// @Tags         Tags
// @Accept       json
// @Produce      json
// @Param        limit   query     int     false  "Max items to return (default 50, max 200)"
// @Param        cursor  query     string  false  "Opaque pagination cursor from a prior next_cursor"
// @Success      200  {object}  TagListResponse
// @Failure      401  {object}  ErrorResponse
// @Failure      500  {object}  ErrorResponse
// @Security     BearerToken
// @Router       /tags [get]
func (h *tagsAPIHandler) List(w http.ResponseWriter, r *http.Request) {
	user := auth.UserFromContext(r.Context())
	if user == nil {
		writeError(w, http.StatusUnauthorized, "unauthorized", "unauthorized")
		return
	}

	// Governing: SPEC-0005 REQ "Pagination" — ?limit (default 50, max 200) + opaque ?cursor
	limit := parseLimit(r)
	cursorName, cursorID := parseCursor(r)

	// Fetch limit+1 to detect whether another page exists.
	tagsWithCounts, err := h.tags.ListWithCountsPaginated(r.Context(), limit+1, cursorName, cursorID)
	if err != nil {
		writeError(w, http.StatusInternalServerError, "internal error", "internal_error")
		return
	}

	var nextCursor *string
	if len(tagsWithCounts) > limit {
		last := tagsWithCounts[limit-1]
		c := encodeCursor(last.Name, last.ID)
		nextCursor = &c
		tagsWithCounts = tagsWithCounts[:limit]
	}

	resp := &TagListResponse{Tags: make([]*TagResponse, 0, len(tagsWithCounts)), NextCursor: nextCursor}
	for _, t := range tagsWithCounts {
		resp.Tags = append(resp.Tags, &TagResponse{
			Slug:      t.Slug,
			Name:      t.Name,
			LinkCount: t.Count,
		})
	}

	writeJSON(w, http.StatusOK, resp)
}

// ListLinks returns links tagged with the given slug.
// GET /api/v1/tags/{slug}/links
// Governing: SPEC-0005 REQ "Tags" — admin sees all links; non-admin sees only owned links.
//
// @Summary      List links by tag
// @Description  Returns links with the given tag. Admins see all; non-admins see only owned links.
// @Tags         Tags
// @Accept       json
// @Produce      json
// @Param        slug  path      string  true  "Tag slug"
// @Success      200   {object}  LinkListResponse
// @Failure      401   {object}  ErrorResponse
// @Failure      404   {object}  ErrorResponse
// @Failure      500   {object}  ErrorResponse
// @Security     BearerToken
// @Router       /tags/{slug}/links [get]
func (h *tagsAPIHandler) ListLinks(w http.ResponseWriter, r *http.Request) {
	user := auth.UserFromContext(r.Context())
	if user == nil {
		writeError(w, http.StatusUnauthorized, "unauthorized", "unauthorized")
		return
	}

	tagSlug := chi.URLParam(r, "slug")

	// Verify the tag exists.
	_, err := h.tags.GetBySlug(r.Context(), tagSlug)
	if errors.Is(err, store.ErrNotFound) {
		writeError(w, http.StatusNotFound, "tag not found", "not_found")
		return
	}
	if err != nil {
		writeError(w, http.StatusInternalServerError, "internal error", "internal_error")
		return
	}

	var links []*store.Link
	if user.IsAdmin() {
		links, err = h.links.ListByTag(r.Context(), tagSlug)
	} else {
		links, err = h.links.ListByOwnerAndTag(r.Context(), user.ID, tagSlug)
	}
	if err != nil {
		writeError(w, http.StatusInternalServerError, "internal error", "internal_error")
		return
	}

	resp := &LinkListResponse{Links: make([]*LinkResponse, 0, len(links))}
	for _, l := range links {
		resp.Links = append(resp.Links, &LinkResponse{
			ID:          l.ID,
			Slug:        l.Slug,
			URL:         l.URL,
			Title:       l.Title,
			Description: l.Description,
			CreatedAt:   l.CreatedAt,
			UpdatedAt:   l.UpdatedAt,
		})
	}

	writeJSON(w, http.StatusOK, resp)
}